refactored Pagination component to extract a custom hook
